requirements.txt
anthropic
chromadb
crewai
crewai_tools
langchain
langchain_ollama
langchain_openai
pandas
pydantic
pysqlite3-binary
python-dotenv
PyYAML
streamlit
streamlit-mermaid
